using exiftool ipc to get jpeg images rather than dcraw in web workers
